Partial fix for #9
This approach is comparing vertices for a common edge between neighboring cells.
Due to clipping cells may share a vertex (same position) which is indexed with different values.
Ideally that needs to be fixed (which has additional benefit of yeild simpler logic here and elsewhere) so the test cases could pass.